refactor(auxiliary-data): host the feature-name-kind flag beside its rule
`FeatureNameKindArg` lived in `senna/src/masked_topic.rs`, so any other crate
needing the same `--feature-name-kind` vocabulary had to duplicate it. The
rule it selects, `FeatureNameKind`, already lives in auxiliary-data; the flag
belongs next to it.

Moved verbatim, serde derives included, and senna re-exports it so its own
paths are unchanged. auxiliary-data gains clap and serde for the derives.
